fos/requirements.txt
boto3==1.13.8
Django==2.2
django-mptt==0.11.0
django-storages==1.9.1
djangorestframework==3.11.0
mysql-connector==2.2.9
mysqlclient==1.4.6
psycopg2==2.8.5
python-memcached==1.59